[심수현]채팅html 합침, 읽음 로직 수정 필요
