+++ src/main/java/com/codenal/chat/repository/ChatReadRepository.java
@@ -11,20 +11,24 @@ import com.codenal.chat.domain.ChatParticipants;
 import com.codenal.chat.domain.ChatRead;
 import com.codenal.chat.domain.ChatRoom;
 
-public interface ChatReadRepository extends JpaRepository<ChatRead, Integer>{
+public interface ChatReadRepository extends JpaRepository<ChatRead, Integer> {
 
-	@Query(value="SELECT r FROM ChatRead r WHERE r.chatMsg = ?1 AND r.chatParticipant = ?2")
+	@Query(value = "SELECT r FROM ChatRead r WHERE r.chatMsg = ?1 AND r.chatParticipant = ?2")
 	List<ChatRead> findAllById(int chatMsg, int participantNo);
 
 	@Modifying
 	@Query(value="UPDATE ChatRead r SET r.isReceiverRead = 'Y' WHERE r.chatParticipant.participantNo = ?1")
 	void updateByRead(int participantNo);
-
 	
-    Long countByChatParticipant_ParticipantNoAndIsReceiverRead(Integer participantNo, Character readStatus);
+	@Modifying
+	@Query("UPDATE ChatRead r SET r.isReceiverRead = 'Y' " 
+			+ "WHERE r.chatParticipant.participantNo IN "
+			+ "(SELECT p.participantNo FROM ChatParticipants p WHERE p.employee.empId = ?1)")
+	void updateByRead(Long empId);
 
-    
-    @Modifying
-    @Query("DELETE FROM ChatRead cr WHERE cr.chatMsg IN (SELECT cm FROM ChatMsg cm WHERE cm.chatRoom = ?1)")
-    void deleteByChatRoom(ChatRoom chatRoom);
+	Long countByChatParticipant_ParticipantNoAndIsReceiverRead(Integer participantNo, Character readStatus);
+
+	@Modifying
+	@Query("DELETE FROM ChatRead cr WHERE cr.chatMsg IN (SELECT cm FROM ChatMsg cm WHERE cm.chatRoom = ?1)")
+	void deleteByChatRoom(ChatRoom chatRoom);
 }
