+++ src/main/resources/application.properties
@@ -2,11 +2,13 @@ server.port=8083
 spring.thymeleaf.prefix=classpath:/templates/
 #spring.thymeleaf.check-template-location=false
 
-# MySQL
-spring.datasource.url=jdbc:mysql://localhost:3306/withxwork
+spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
+
+spring.datasource.url=jdbc:mariadb://127.0.0.1:3306/withxwork
+
 spring.datasource.username=codenal
 spring.datasource.password=1234
-
-# Hibernate JPA
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MariaDBDialect
+spring.application.name=codenal
+spring.servlet.multipart.max-file-size=500MB
+spring.servlet.multipart.max-request-size=500MB
